Renamed geometry types to make better sense.
