+++ lib/models/geometry/geometry_polygon.dart
@@ -3,9 +3,9 @@ import 'dart:convert';
 import 'package:mapbox_gl/mapbox_gl.dart';
 import 'package:mapbox_gl/models/bounding_box.dart';
 import 'package:mapbox_gl/models/geometry/geometry.dart';
-import 'package:mapbox_gl/models/geometry/point.dart';
+import 'package:mapbox_gl/models/geometry/geometry_point.dart';
 
-class Polygon extends Geometry {
+class GeometryPolygon extends Geometry {
   static const TYPE = "Polygon";
 
   String type;
@@ -14,17 +14,17 @@ class Polygon extends Geometry {
 
   final BoundingBox bbox;
 
-  Polygon(this.coordinates, {this.bbox}) {
+  GeometryPolygon(this.coordinates, {this.bbox}) {
     type = TYPE;
   }
 
-  static Polygon fromJson(String jsonString) =>
+  static GeometryPolygon fromJson(String jsonString) =>
       fromMap(json.decode(jsonString));
 
-  static Polygon fromMap(Map<String, dynamic> map) {
+  static GeometryPolygon fromMap(Map<String, dynamic> map) {
     if (!map.containsKey('coordinates')) return null;
 
-    return Polygon(
+    return GeometryPolygon(
         (map['coordinates'] as List)
             .map((a) =>
                 (a as List).map((b) => GeometryPoint.fromMap(b)).toList())
